Extracted XCode project generation to separate module
